+++ tests/fixtures/load_services.py
@@ -11,6 +11,7 @@ from typing import (
 )
 
 import pytest
+from tenacity import wait_fixed
 
 if TYPE_CHECKING:
     from collections.abc import Coroutine, Generator, Sequence
@@ -56,14 +57,29 @@ def load_services(wait_for: WaitFor) -> Generator[LoadServices, None, None]:
         from ubo_app.service_thread import load_services
 
         load_services(service_ids, gap_duration=gap_duration)
+        ids = list(service_ids)
 
-        @wait_for(run_async=cast(Literal[True], run_async), timeout=timeout)
+        @wait_for(
+            run_async=cast(Literal[True], run_async),
+            timeout=timeout,
+            wait=wait_fixed(1),
+        )
         def check() -> None:
-            for service_id in service_ids:
+            for service_id in list(ids):
+                assert any(
+                    service.service_id == service_id
+                    for service in REGISTERED_PATHS.values()
+                ), f'{service_id} not loaded'
                 assert any(
                     service.service_id == service_id and service.is_alive()
                     for service in REGISTERED_PATHS.values()
-                ), f'{service_id} not loaded'
+                ), f'{service_id} not alive'
+                assert any(
+                    service.service_id == service_id and service.is_started
+                    for service in REGISTERED_PATHS.values()
+                ), f'{service_id} not started'
+
+                ids.remove(service_id)
 
         return check()