+++ ubo_app/logger.py
@@ -1,7 +1,6 @@
 # ruff: noqa: D100, D101, D102, D103, D104, D107
 from __future__ import annotations
 
-import atexit
 import json
 import logging
 import logging.handlers
@@ -11,6 +10,8 @@ from typing import TYPE_CHECKING, ClassVar, cast
 
 if TYPE_CHECKING:
     from collections.abc import Mapping
+
+    from ubo_app.utils.types import Subscriptions
 
 VERBOSE = 5
 
@@ -156,7 +157,10 @@ class StdOutExtraFormatter(ExtraFormatter):
         return f'{color}{string}{self.RESET}'
 
 
-def add_stdout_handler(logger: UboLogger, level: int = logging.DEBUG) -> None:
+def add_stdout_handler(
+    logger: UboLogger,
+    level: int = logging.DEBUG,
+) -> Subscriptions:
     stdout_handler = logging.StreamHandler(sys.stdout)
     stdout_handler.setLevel(level)
     formatter = StdOutExtraFormatter(
@@ -164,14 +168,22 @@ def add_stdout_handler(logger: UboLogger, level: int = logging.DEBUG) -> None:
         '%Y-%m-%d %H:%M:%S',
     )
     if level <= logging.DEBUG:
-        formatter.max_length = 400
+        formatter.max_length = 1000
     stdout_handler.setFormatter(formatter)
     logger.addHandler(stdout_handler)
 
-    atexit.register(stdout_handler.flush)
+    def cleanup() -> None:
+        stdout_handler.flush()
+        stdout_handler.close()
+        logger.removeHandler(stdout_handler)
+
+    return [cleanup]
 
 
-def add_file_handler(logger: UboLogger, level: int = logging.DEBUG) -> None:
+def add_file_handler(
+    logger: UboLogger,
+    level: int = logging.DEBUG,
+) -> Subscriptions:
     file_handler = logging.handlers.RotatingFileHandler(
         f'{logger.name}.log',
         backupCount=3,
@@ -186,7 +198,12 @@ def add_file_handler(logger: UboLogger, level: int = logging.DEBUG) -> None:
     )
     logger.addHandler(file_handler)
 
-    atexit.register(file_handler.flush)
+    def cleanup() -> None:
+        file_handler.flush()
+        file_handler.close()
+        logger.removeHandler(file_handler)
+
+    return [cleanup]
 
 
 def get_log_level() -> int | None:
@@ -218,13 +235,14 @@ def get_gui_log_level() -> int | None:
     return None
 
 
-def setup_logging() -> None:
+def setup_loggers() -> Subscriptions:
     level = get_log_level()
+    subscriptions: Subscriptions = []
 
     if level is not None:
         logger.setLevel(level)
-        add_file_handler(logger, level)
-        add_stdout_handler(logger, level)
+        subscriptions.extend(add_file_handler(logger, level))
+        subscriptions.extend(add_stdout_handler(logger, level))
 
     gui_level = get_gui_log_level()
 
@@ -235,5 +253,7 @@ def setup_logging() -> None:
         ubo_gui.logger.add_file_handler(gui_level)
         ubo_gui.logger.add_stdout_handler(gui_level)
 
+    return subscriptions
 
-__all__ = ('add_file_handler', 'add_stdout_handler', 'logger', 'setup_logging')
+
+__all__ = ('add_file_handler', 'add_stdout_handler', 'logger', 'setup_loggers')