+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 - refactor: use the new subscriptions return value of the `setup` function for different services to report subscriptions so that the service manager can clean up the subscriptions when the service is stopped
 - fix(core): remove the reference of the async task handle after it is done to avoid memory leaks
 - feat(core): clean up status icons of a service when it stops
+- refactor(tests): make different parts of code explicitly return their cleanup functions and make test_services wait until all services are completely loaded without any errors plus small improvements to logs
 
 ## Version 1.2.2